backend/requirements.txt
numpy
pandas
pyodbc
requests
scikit-learn
seaborn
tensorflow
torch
matplotlib
transformers
joblib
openrouteservice
fastapi
uvicorn
python-dotenv
sqlalchemy
asyncpg
urllib
pydantic